added decompression support to ANFO file parser (currently gzip and plain are supported)
git-svn-id: http://biofs04/svn/bioinf/anfo/trunk@1609 93d9a472-daf7-0310-8f21-d7befa554562

+++ anfo2sam.cc
@@ -84,7 +84,7 @@ int protoHit_2_bam_Hit(output::Result &result){
 /*ISIZE*/   std::cout << "0" << "\t";
 
 /*SEQ*/     std::cout << result.sequence() << "\t";
-            for (int i = 0; i < result.quality().size(); i++)
+            for (size_t i = 0; i < result.quality().size(); i++)
 /*QUAL*/        std::cout << char((uint8_t)result.quality()[i] + 33);
 
 /*[TAGS]*/
